+++ src/main/main_preload.js
@@ -1,0 +1,5 @@
+const { ipcRenderer, clipboard, webFrame } = require('electron');
+
+window.ipcRenderer = ipcRenderer;
+window.clipboard = clipboard;
+window.webFrame = webFrame;