+++ src/renderer/shell/actions/nav.js
@@ -1,8 +1,7 @@
 import {
   UPDATE_TARGET_URL,
-  UPDATE_LOADING,
-  UPDATE_CAN_GO_BACK,
-  UPDATE_CAN_GO_FORWARD,
+  UPDATE_IS_FAILED,
+  UPDATE_IS_LOADING,
 } from '../constants/actions';
 
 export const updateTargetUrl = targetUrl => ({
@@ -10,17 +9,12 @@ export const updateTargetUrl = targetUrl => ({
   targetUrl,
 });
 
-export const updateLoading = isLoading => ({
-  type: UPDATE_LOADING,
+export const updateIsLoading = isLoading => ({
+  type: UPDATE_IS_LOADING,
   isLoading,
 });
 
-export const updateCanGoBack = canGoBack => ({
-  type: UPDATE_CAN_GO_BACK,
-  canGoBack,
-});
-
-export const updateCanGoForward = canGoForward => ({
-  type: UPDATE_CAN_GO_FORWARD,
-  canGoForward,
+export const updateIsFailed = isFailed => ({
+  type: UPDATE_IS_FAILED,
+  isFailed,
 });