+++ src/renderer/shell/actions/settings.js
@@ -1,7 +1,4 @@
-import { ipcRenderer } from 'electron';
-
-import { TOGGLE_SETTING_DIALOG, SET_BEHAVIOR, SET_BEHAVIORS } from '../constants/actions';
-import defaultSettings from '../constants/defaultSettings';
+import { TOGGLE_SETTING_DIALOG, SET_BEHAVIOR } from '../constants/actions';
 
 export const toggleSettingDialog = () => ({
   type: TOGGLE_SETTING_DIALOG,
@@ -15,12 +12,4 @@ export const setBehavior = (name, val) => (dispatch) => {
   });
 
   ipcRenderer.send('set-setting', `behaviors.${window.shellInfo.id}.${name}`, val);
-};
-
-export const getBehaviors = () => (dispatch) => {
-  const behaviors = ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}`, defaultSettings);
-  dispatch({
-    type: SET_BEHAVIORS,
-    behaviors,
-  });
 };