+++ src/renderer/shell/components/App.js
@@ -1,28 +1,47 @@
 /* eslint-disable no-console */
-import { ipcRenderer, clipboard } from 'electron';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { NonIdealState, Button, Intent, Classes, Hotkey, Hotkeys, HotkeysTarget } from '@blueprintjs/core';
+
+import Immutable from 'immutable';
 
 import extractDomain from '../libs/extractDomain';
-import { updateTargetUrl, updateLoading, updateCanGoBack, updateCanGoForward } from '../actions/nav';
-import { toggleSettingDialog, getBehaviors } from '../actions/settings';
+import {
+  updateTargetUrl,
+  updateIsFailed,
+  updateIsLoading,
+} from '../actions/nav';
+import { toggleSettingDialog } from '../actions/settings';
 import { toggleFindInPageDialog, updateFindInPageMatches } from '../actions/findInPage';
 import { screenResize } from '../actions/screen';
+import {
+  addTab,
+  closeTab,
+  updateActiveTab,
+  updateTabLastURL,
+  updateCanGoBack,
+  updateCanGoForward,
+} from '../actions/tabs';
 
+import LeftNav from './LeftNav';
 import WebView from './WebView';
 import Settings from './Settings';
 import Nav from './Nav';
 import FindInPage from './FindInPage';
 import showUpdateToast from '../../shared/components/showUpdateToast';
 
+@HotkeysTarget
 class App extends React.Component {
   constructor() {
     super();
     this.handleNewWindow = this.handleNewWindow.bind(this);
+    this.handleDidFailLoad = this.handleDidFailLoad.bind(this);
     this.handleDidStopLoading = this.handleDidStopLoading.bind(this);
     this.handleDidGetRedirectRequest = this.handleDidGetRedirectRequest.bind(this);
     this.handleUpdateTargetUrl = this.handleUpdateTargetUrl.bind(this);
+
+    this.webViews = {};
   }
 
   componentDidMount() {
@@ -30,18 +49,17 @@ class App extends React.Component {
       requestToggleSettingDialog,
       requestToggleFindInPageDialog,
       requestUpdateFindInPageMatches,
-      requestGetBehaviors,
+      requestAddTab,
+      requestCloseTab,
       onResize,
     } = this.props;
-    const c = this.c;
-
-    requestGetBehaviors();
 
     showUpdateToast();
 
     window.addEventListener('resize', onResize);
 
     ipcRenderer.on('toggle-dev-tools', () => {
+      const c = this.webViews[this.props.activeTabId];
       c.openDevTools();
     });
 
@@ -51,6 +69,7 @@ class App extends React.Component {
 
     ipcRenderer.on('toggle-find-in-page-dialog', () => {
       if (this.props.findInPageIsOpen) {
+        const c = this.webViews[this.props.activeTabId];
         c.stopFindInPage('clearSelection');
         requestUpdateFindInPageMatches(0, 0);
       }
@@ -58,42 +77,74 @@ class App extends React.Component {
     });
 
     ipcRenderer.on('change-zoom', (event, message) => {
+      const c = this.webViews[this.props.activeTabId];
       c.setZoomFactor(message);
     });
 
     ipcRenderer.on('reload', () => {
+      const c = this.webViews[this.props.activeTabId];
       c.reload();
     });
 
     ipcRenderer.on('go-back', () => {
+      const c = this.webViews[this.props.activeTabId];
       c.goBack();
     });
 
     ipcRenderer.on('go-forward', () => {
+      const c = this.webViews[this.props.activeTabId];
       c.goForward();
     });
 
     ipcRenderer.on('go-home', () => {
+      const c = this.webViews[this.props.activeTabId];
       c.loadURL(this.props.customHome || window.shellInfo.url);
     });
 
     ipcRenderer.on('go-to-url', (e, url) => {
+      const c = this.webViews[this.props.activeTabId];
       c.loadURL(url);
     });
 
     ipcRenderer.on('copy-url', () => {
+      const c = this.webViews[this.props.activeTabId];
       const currentURL = c.getURL();
       clipboard.writeText(currentURL);
     });
+
+    ipcRenderer.on('new-tab', () => {
+      requestAddTab();
+    });
+
+    ipcRenderer.on('close-tab', () => {
+      requestCloseTab(this.props.activeTabId);
+    });
   }
 
-  componentDidUpdate() {
-    const { findInPageIsOpen, findInPageText } = this.props;
-    const c = this.c;
+  componentDidUpdate(prevProps) {
+    const {
+      activeTabId,
+      findInPageIsOpen,
+      findInPageText,
+    } = this.props;
+
+    const c = this.webViews[activeTabId];
 
     // Restart search if text is available
     if (findInPageIsOpen && findInPageText.length > 0) {
       c.findInPage(findInPageText, { forward: true });
+    }
+
+    // https://github.com/electron/electron/issues/5110#issuecomment-208991872
+    if (prevProps.activeTabId !== activeTabId) {
+      if (this.webViews[prevProps.activeTabId]) {
+        this.webViews[prevProps.activeTabId].c.style.flex = '0 1';
+        this.webViews[prevProps.activeTabId].c.style.width = '0';
+        this.webViews[prevProps.activeTabId].c.style.height = '0';
+      }
+      this.webViews[activeTabId].c.style.flex = null;
+      this.webViews[activeTabId].c.style.width = '100%';
+      this.webViews[activeTabId].c.style.height = '100%';
     }
   }
 
@@ -102,7 +153,12 @@ class App extends React.Component {
   }
 
   handleDidGetRedirectRequest(e) {
-    const c = this.c;
+    const {
+      activeTabId,
+    } = this.props;
+
+    const c = this.webViews[activeTabId];
+
     const { newURL, isMainFrame } = e;
     // https://github.com/webcatalog/webcatalog/issues/42
     if (isMainFrame && extractDomain(newURL) === 'twitter.com') {
@@ -112,14 +168,16 @@ class App extends React.Component {
   }
 
   handleNewWindow(e) {
+    const { activeTabId } = this.props;
+
     const nextUrl = e.url;
-    const c = this.c;
+
+    const c = this.webViews[activeTabId];
+
     console.log(`newWindow: ${nextUrl}`);
     // open external url in browser if domain doesn't match.
     const curDomain = extractDomain(window.shellInfo.url);
     const nextDomain = extractDomain(nextUrl);
-
-    console.log(nextDomain);
 
     // open new window
     if (
@@ -143,17 +201,31 @@ class App extends React.Component {
     ipcRenderer.send('open-in-browser', nextUrl);
   }
 
+  handleDidFailLoad(e) {
+    // errorCode -3: cancelling
+    console.log('Error: ', e);
+    if (e.isMainFrame && e.errorCode < 0 && e.errorCode !== -3) {
+      const { requestUpdateIsFailed } = this.props;
+      requestUpdateIsFailed(true);
+    }
+  }
+
   handleDidStopLoading() {
     const {
-      requestUpdateLoading, requestUpdateCanGoBack, requestUpdateCanGoForward,
+      activeTabId,
+      requestUpdateIsLoading,
+      requestUpdateCanGoBack,
+      requestUpdateCanGoForward,
+      requestSetTabLastUrl,
     } = this.props;
-    const c = this.c;
 
-    requestUpdateLoading(false);
-    requestUpdateCanGoBack(c.canGoBack());
-    requestUpdateCanGoForward(c.canGoForward());
+    const c = this.webViews[activeTabId];
 
-    ipcRenderer.send('set-setting', `lastPages.${window.shellInfo.id}`, c.getURL());
+    requestUpdateIsLoading(false);
+    requestUpdateCanGoBack(activeTabId, c.canGoBack());
+    requestUpdateCanGoForward(activeTabId, c.canGoForward());
+
+    requestSetTabLastUrl(activeTabId, c.getURL());
   }
 
   handlePageTitleUpdated({ title }) {
@@ -171,14 +243,43 @@ class App extends React.Component {
     requestUpdateTargetUrl(url);
   }
 
+  renderHotkeys() {
+    const {
+      tabs,
+      requestSetActiveTab,
+    } = this.props;
+
+    return (
+      <Hotkeys>
+        {tabs.map((tab, tabIndex) => (
+          <Hotkey
+            key={tab.get('id')}
+            global
+            combo={`mod + ${tabIndex + 1}`}
+            label={`Tab ${tabIndex + 1}`}
+            onKeyDown={() => requestSetActiveTab(tab.get('id'))}
+          />
+        ))}
+      </Hotkeys>
+    );
+  }
+
   render() {
     const {
-      url, findInPageIsOpen, isFullScreen, customHome,
-      requestUpdateLoading, requestUpdateFindInPageMatches,
+      findInPageIsOpen,
+      isFailed,
+      isFullScreen,
+      customHome,
+      rememberLastPage,
+      tabs,
+      activeTabId,
+      requestUpdateIsFailed,
+      requestUpdateIsLoading,
+      requestUpdateFindInPageMatches,
       targetUrl,
     } = this.props;
 
-    const showNav = process.platform === 'darwin' && !isFullScreen;
+    const showNav = process.env.PLATFORM === 'darwin' && !isFullScreen;
 
     let usedHeight = showNav ? 32 : 0;
     if (findInPageIsOpen) usedHeight += 50;
@@ -189,47 +290,104 @@ class App extends React.Component {
           height: '100vh',
         }}
       >
+        {isFailed ? (
+          <div
+            style={{
+              height: '100vh',
+              width: '100vw',
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+            }}
+          >
+            <NonIdealState
+              visual="error"
+              className="no-connection"
+              title="Internet Connection"
+              description="Please check your Internet connection and try again."
+              action={(
+                <Button
+                  iconName="repeat"
+                  intent={Intent.PRIMARY}
+                  className={Classes.LARGE}
+                  text="Try Again"
+                  onClick={() => {
+                    requestUpdateIsFailed(false);
+                    this.webViews.reload();
+                  }}
+                />
+              )}
+            />
+          </div>
+        ) : null}
         {showNav ? (
           <Nav
-            onHomeButtonClick={() => this.c.loadURL(customHome || window.shellInfo.url)}
-            onBackButtonClick={() => this.c.goBack()}
-            onForwardButtonClick={() => this.c.goForward()}
-            onRefreshButtonClick={() => this.c.reload()}
+            onHomeButtonClick={() => {
+              const c = this.webViews[activeTabId];
+              c.loadURL(customHome || window.shellInfo.url);
+            }}
+            onBackButtonClick={() => {
+              const c = this.webViews[activeTabId];
+              c.goBack();
+            }}
+            onForwardButtonClick={() => {
+              const c = this.webViews[activeTabId];
+              c.goForward();
+            }}
+            onRefreshButtonClick={() => {
+              const c = this.webViews[activeTabId];
+              c.reload();
+            }}
           />
         ) : null}
         {findInPageIsOpen ? (
           <FindInPage
-            onRequestFind={(text, forward) => this.c.findInPage(text, { forward })}
+            onRequestFind={(text, forward) => {
+              const c = this.webViews[activeTabId];
+              c.findInPage(text, { forward });
+            }}
             onRequestStopFind={() => {
-              this.c.stopFindInPage('clearSelection');
+              const c = this.webViews[activeTabId];
+              c.stopFindInPage('clearSelection');
               requestUpdateFindInPageMatches(0, 0);
             }}
           />
         ) : null}
-        <div style={{ height: `calc(100vh - ${usedHeight}px)`, width: '100%' }}>
-          <WebView
-            ref={(c) => { this.c = c; }}
-            src={url}
-            style={{ height: '100%', width: '100%' }}
-            className="webview"
-            plugins
-            allowpopups
-            autoresize
-            preload="../preload.js"
-            // enable nodeintegration in testing mode (mainly for Spectron)
-            nodeintegration={window.shellInfo.isTesting}
-            useragent={window.shellInfo.userAgent}
-            partition={`persist:${window.shellInfo.id}`}
-            onDidGetRedirectRequest={this.handleDidGetRedirectRequest}
-            onNewWindow={this.handleNewWindow}
-            onDidStartLoading={() => requestUpdateLoading(true)}
-            onDidStopLoading={this.handleDidStopLoading}
-            onFoundInPage={({ result }) => {
-              requestUpdateFindInPageMatches(result.activeMatchOrdinal, result.matches);
-            }}
-            onPageTitleUpdated={this.handlePageTitleUpdated}
-            onUpdateTargetUrl={this.handleUpdateTargetUrl}
-          />
+        <div style={{ height: `calc(100vh - ${usedHeight}px)`, width: '100vw', display: 'flex' }}>
+          {tabs.size > 1 ? <LeftNav /> : null}
+          <div style={{ height: `calc(100vh - ${usedHeight}px)`, width: '100%' }}>
+            {tabs.map(tab => (
+              <WebView
+                key={tab.get('id')}
+                ref={(c) => { this.webViews[tab.get('id')] = c; }}
+                src={rememberLastPage ? (tab.get('lastUrl') || customHome || window.shellInfo.url) : (customHome || window.shellInfo.url)}
+                style={{
+                  height: !tab.get('isActive') ? 0 : '100%',
+                  width: !tab.get('isActive') ? 0 : '100%',
+                  flex: !tab.get('isActive') ? '0 1' : null,
+                }}
+                className="webview"
+                plugins
+                allowpopups
+                autoresize
+                preload={window.shellInfo.preload}
+                // enable nodeintegration in testing mode (mainly for Spectron)
+                nodeintegration={window.shellInfo.isTesting}
+                useragent={window.shellInfo.userAgent}
+                partition={tab.get('partition')}
+                onDidGetRedirectRequest={this.handleDidGetRedirectRequest}
+                onDidFailLoad={this.handleDidFailLoad}
+                onNewWindow={this.handleNewWindow}
+                onDidStartLoading={() => requestUpdateIsLoading(true)}
+                onDidStopLoading={this.handleDidStopLoading}
+                onFoundInPage={({ result }) => {
+                  requestUpdateFindInPageMatches(result.activeMatchOrdinal, result.matches);
+                }}
+                onPageTitleUpdated={this.handlePageTitleUpdated}
+                onUpdateTargetUrl={this.handleUpdateTargetUrl}
+              />
+            ))}
+          </div>
         </div>
         <div
           style={{
@@ -257,42 +415,62 @@ class App extends React.Component {
 }
 
 App.propTypes = {
-  url: PropTypes.string.isRequired,
   findInPageIsOpen: PropTypes.bool.isRequired,
   findInPageText: PropTypes.string.isRequired,
   isFullScreen: PropTypes.bool,
+  isFailed: PropTypes.bool,
   customHome: PropTypes.string,
+  rememberLastPage: PropTypes.bool,
   targetUrl: PropTypes.string,
+  tabs: PropTypes.instanceOf(Immutable.List),
+  activeTabId: PropTypes.string,
   onResize: PropTypes.func.isRequired,
   requestUpdateTargetUrl: PropTypes.func.isRequired,
-  requestUpdateLoading: PropTypes.func.isRequired,
+  requestUpdateIsFailed: PropTypes.func.isRequired,
+  requestUpdateIsLoading: PropTypes.func.isRequired,
   requestUpdateCanGoBack: PropTypes.func.isRequired,
   requestUpdateCanGoForward: PropTypes.func.isRequired,
   requestToggleSettingDialog: PropTypes.func.isRequired,
   requestToggleFindInPageDialog: PropTypes.func.isRequired,
   requestUpdateFindInPageMatches: PropTypes.func.isRequired,
-  requestGetBehaviors: PropTypes.func.isRequired,
+  requestAddTab: PropTypes.func.isRequired,
+  requestCloseTab: PropTypes.func.isRequired,
+  requestSetActiveTab: PropTypes.func.isRequired,
+  requestSetTabLastUrl: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = state => ({
-  findInPageIsOpen: state.findInPage.get('isOpen'),
-  findInPageText: state.findInPage.get('text'),
-  isFullScreen: state.screen.get('isFullScreen'),
-  customHome: state.settings.getIn(['behaviors', 'customHome']),
-  targetUrl: state.nav.get('targetUrl'),
-});
+const mapStateToProps = (state) => {
+  const tabs = state.tabs.get('list');
+  const activeTab = tabs.find(tab => tab.get('isActive'));
+
+  return {
+    findInPageIsOpen: state.findInPage.get('isOpen'),
+    findInPageText: state.findInPage.get('text'),
+    isFullScreen: state.screen.get('isFullScreen'),
+    isFailed: state.nav.get('isFailed'),
+    customHome: state.settings.get('customHome'),
+    rememberLastPage: state.settings.get('rememberLastPage'),
+    targetUrl: state.nav.get('targetUrl'),
+    tabs,
+    activeTabId: activeTab.get('id'),
+  };
+};
 
 const mapDispatchToProps = dispatch => ({
   onResize: () => dispatch(screenResize(window.innerWidth)),
   requestUpdateTargetUrl: targetUrl => dispatch(updateTargetUrl(targetUrl)),
-  requestUpdateLoading: isLoading => dispatch(updateLoading(isLoading)),
-  requestUpdateCanGoBack: canGoBack => dispatch(updateCanGoBack(canGoBack)),
-  requestUpdateCanGoForward: canGoForward => dispatch(updateCanGoForward(canGoForward)),
+  requestUpdateIsFailed: isFailed => dispatch(updateIsFailed(isFailed)),
+  requestUpdateIsLoading: isLoading => dispatch(updateIsLoading(isLoading)),
+  requestUpdateCanGoBack: (tabId, canGoBack) => dispatch(updateCanGoBack(tabId, canGoBack)),
+  requestUpdateCanGoForward: (tabId, canGoForward) => dispatch(updateCanGoForward(tabId, canGoForward)),
   requestToggleSettingDialog: () => dispatch(toggleSettingDialog()),
   requestToggleFindInPageDialog: () => dispatch(toggleFindInPageDialog()),
   requestUpdateFindInPageMatches: (activeMatch, matches) =>
     dispatch(updateFindInPageMatches(activeMatch, matches)),
-  requestGetBehaviors: () => dispatch(getBehaviors()),
+  requestAddTab: () => dispatch(addTab()),
+  requestCloseTab: tabId => dispatch(closeTab(tabId)),
+  requestSetActiveTab: isActive => dispatch(updateActiveTab(isActive)),
+  requestSetTabLastUrl: (tabId, lastUrl) => dispatch(updateTabLastURL(tabId, lastUrl)),
 });
 
 export default connect(