+++ src/renderer/shell/reducers/index.js
@@ -4,12 +4,14 @@ import nav from './nav';
 import settings from './settings';
 import findInPage from './findInPage';
 import screen from './screen';
+import tabs from './tabs';
 
 const rootReducer = combineReducers({
   nav,
   settings,
   findInPage,
   screen,
+  tabs,
 });
 
 export default rootReducer;