+++ src/renderer/shell/reducers/settings.js
@@ -1,13 +1,19 @@
 import Immutable from 'immutable';
 
-import { TOGGLE_SETTING_DIALOG, SET_BEHAVIOR, SET_BEHAVIORS } from '../constants/actions';
+import { TOGGLE_SETTING_DIALOG, SET_BEHAVIOR } from '../constants/actions';
 import defaultSettings from '../constants/defaultSettings';
 
 const initialState = Immutable.fromJS({
   isOpen: false,
-  behaviors: defaultSettings,
+  swipeToNavigate: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.swipeToNavigate`, defaultSettings.swipeToNavigate),
+  rememberLastPage: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.rememberLastPage`, defaultSettings.rememberLastPage),
+  quitOnLastWindow: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.quitOnLastWindow`, defaultSettings.quitOnLastWindow),
+  blockAds: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.blockAds`, defaultSettings.blockAds),
+  customHome: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.customHome`, defaultSettings.customHome),
+  injectedCSS: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.injectedCSS`, defaultSettings.injectedCSS),
+  injectedJS: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.injectedJS`, defaultSettings.injectedJS),
+  customUserAgent: ipcRenderer.sendSync('get-setting', `behaviors.${window.shellInfo.id}.customUserAgent`, defaultSettings.customUserAgent),
 });
-
 
 const settings = (state = initialState, action) => {
   switch (action.type) {
@@ -15,10 +21,7 @@ const settings = (state = initialState, action) => {
       return state.set('isOpen', !state.get('isOpen'));
     }
     case SET_BEHAVIOR: {
-      return state.setIn(['behaviors', action.behaviorName], action.behaviorVal);
-    }
-    case SET_BEHAVIORS: {
-      return state.set('behaviors', Immutable.Map(action.behaviors));
+      return state.set(action.behaviorName, action.behaviorVal);
     }
     default:
       return state;