+++ src/renderer/store/index.js
@@ -1,4 +1,3 @@
-import { ipcRenderer, webFrame } from 'electron';
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';